Dynamic Packet Length Formatting
This commit starts the work to format the initial DWORD values
differently based upon the payload length code.

This has highlighted a considerable need to restructure the way that the
frame is formatted. Each BYTE cell needs to be formatted differently
based upon the payload length code (short, medium, long) and whether the
payload data is masked.

A refactor and restructure is upcoming to address this need before going
further with updating the formatter.